refactor: move torch/assembler.py's schema helpers to _private/schema, matching internal
reorder_output_schema and normalize_scalar_shapes were inline, underscore-
prefixed functions in torch/assembler.py. Internal keeps the real
counterpart of the first one in a dedicated location:
uber-one:uber/ai/michelangelo/sdk/workflow/tasks/tabular_assembler/_private/schema/output_schema.py.
Task-private helpers belong under _private/, not underscore-prefixed
inline in the public assembler module -- this mirrors the existing
_private/model_class/resolve.py pattern already used elsewhere in this
task.

- _private/schema/output_schema.py: reorder_output_schema, moved to match
  internal's exact path and name. Ported internal's fuller test suite
  (duplicate-field-order, empty-schema, object-identity assertions) that
  OSS's version lacked.
- _private/schema/scalar_shapes.py: normalize_scalar_shapes. No internal
  equivalent (see the prior commit for why) -- kept in a separate file from
  output_schema.py rather than merged into it, since only one of the two
  functions actually mirrors an internal module.

torch/assembler.py now imports both from _private.schema; the underscore
prefixes are dropped since privacy comes from the _private/ package
location now, not per-symbol naming, matching the model_class/resolve.py
convention.
