+++ python/michelangelo/workflow/tasks/tabular_assembler/_private/schema/__init__.py
@@ -1,0 +1,8 @@
+"""Schema/sample-data normalization helpers for the tabular assembler."""
+
+from __future__ import annotations
+
+from .output_schema import reorder_output_schema
+from .scalar_shapes import normalize_scalar_shapes
+
+__all__ = ["normalize_scalar_shapes", "reorder_output_schema"]
